requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.7.0
pydantic-settings>=2.4.0
sqlalchemy>=2.0.30
alembic>=1.13.2
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
bcrypt<4.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.1
structlog>=24.1.0
